spli dialog contents into separate composables
