+++ app/src/main/java/com/babylon/wallet/android/presentation/status/transaction/CompletingContent.kt
@@ -1,0 +1,57 @@
+package com.babylon.wallet.android.presentation.status.transaction
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.babylon.wallet.android.R
+import com.babylon.wallet.android.designsystem.theme.RadixTheme
+import com.babylon.wallet.android.designsystem.theme.RadixWalletTheme
+
+@Composable
+internal fun CompletingContent(
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier
+            .fillMaxWidth()
+            .background(color = RadixTheme.colors.defaultBackground)
+            .padding(RadixTheme.dimensions.paddingLarge),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(RadixTheme.dimensions.paddingDefault)
+    ) {
+        Image(
+            painter = painterResource(
+                id = com.babylon.wallet.android.designsystem.R.drawable.check_circle_outline
+            ),
+            alpha = 0.2F,
+            contentDescription = null
+        )
+        Text(
+            text = stringResource(R.string.transactionStatus_completing_text),
+            style = RadixTheme.typography.body1Regular,
+            color = RadixTheme.colors.gray1
+        )
+        Spacer(Modifier.height(36.dp))
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun CompletingBottomDialogPreview() {
+    RadixWalletTheme {
+        CompletingContent()
+    }
+}
