+++ app/src/main/java/com/babylon/wallet/android/presentation/status/transaction/TransactionStatusDialog.kt
@@ -3,42 +3,18 @@ package com.babylon.wallet.android.presentation.status.transaction
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.babylon.wallet.android.R
-import com.babylon.wallet.android.designsystem.theme.RadixTheme
-import com.babylon.wallet.android.designsystem.theme.RadixWalletTheme
 import com.babylon.wallet.android.presentation.ui.composables.BasicPromptAlertDialog
 import com.babylon.wallet.android.presentation.ui.composables.BottomSheetDialogWrapper
-import com.babylon.wallet.android.presentation.ui.composables.FailureDialogContent
-import com.babylon.wallet.android.presentation.ui.composables.actionableaddress.ActionableAddressView
 import com.radixdlt.sargon.DappWalletInteractionErrorType
-import com.radixdlt.sargon.IntentHash
-import com.radixdlt.sargon.extensions.init
 
 @Composable
 @Suppress("CyclomaticComplexMethod")
@@ -102,7 +78,7 @@ fun TransactionStatusDialog(
                     FailureDialogContent(
                         title = title,
                         subtitle = state.failureError,
-                        transactionAddress = state.transactionId,
+                        transactionId = state.transactionId,
                         isMobileConnect = state.status.isMobileConnect
                     )
                 }
@@ -114,7 +90,7 @@ fun TransactionStatusDialog(
                 ) {
                     // Need to send the correct transaction id
                     SuccessContent(
-                        transactionAddress = state.transactionId,
+                        transactionId = state.transactionId,
                         isMobileConnect = state.status.isMobileConnect,
                         dAppName = state.status.dAppName,
                         isInternal = state.status.isInternal
@@ -140,142 +116,4 @@ fun TransactionStatusDialog(
             }
         }
     )
-}
-
-@Composable
-private fun SuccessContent(
-    modifier: Modifier = Modifier,
-    transactionAddress: String,
-    isMobileConnect: Boolean,
-    isInternal: Boolean,
-    dAppName: String?
-) {
-    Column {
-        Column(
-            modifier
-                .fillMaxWidth()
-                .background(color = RadixTheme.colors.defaultBackground)
-                .padding(RadixTheme.dimensions.paddingLarge),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(RadixTheme.dimensions.paddingDefault)
-        ) {
-            Image(
-                painter = painterResource(
-                    id = com.babylon.wallet.android.designsystem.R.drawable.check_circle_outline
-                ),
-                contentDescription = null
-            )
-            Text(
-                text = stringResource(id = R.string.transactionStatus_success_title),
-                style = RadixTheme.typography.title,
-                color = RadixTheme.colors.gray1,
-                textAlign = TextAlign.Center
-            )
-
-            val subtitle = if (isInternal) {
-                stringResource(R.string.transactionStatus_success_text)
-            } else {
-                stringResource(
-                    id = R.string.dAppRequest_completion_subtitle,
-                    dAppName.orEmpty().ifEmpty {
-                        stringResource(id = R.string.dAppRequest_metadata_unknownName)
-                    }
-                )
-            }
-            Text(
-                text = subtitle,
-                style = RadixTheme.typography.body1Regular,
-                color = RadixTheme.colors.gray1,
-                textAlign = TextAlign.Center
-            )
-
-            if (transactionAddress.isNotEmpty()) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.transactionStatus_transactionID_text),
-                        style = RadixTheme.typography.body1Header,
-                        color = RadixTheme.colors.gray1
-                    )
-                    Spacer(modifier = Modifier.width(RadixTheme.dimensions.paddingXSmall))
-                    ActionableAddressView(
-                        transactionId = IntentHash.init(transactionAddress),
-                        textStyle = RadixTheme.typography.body1HighImportance,
-                        textColor = RadixTheme.colors.blue1,
-                        iconColor = RadixTheme.colors.gray2
-                    )
-                }
-            }
-        }
-        if (isMobileConnect) {
-            HorizontalDivider(color = RadixTheme.colors.gray4)
-            Text(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(color = RadixTheme.colors.gray5)
-                    .padding(vertical = RadixTheme.dimensions.paddingLarge, horizontal = RadixTheme.dimensions.paddingXLarge),
-                text = stringResource(id = R.string.mobileConnect_interactionSuccess),
-                style = RadixTheme.typography.body1Regular,
-                color = RadixTheme.colors.gray1,
-                textAlign = TextAlign.Center
-            )
-        }
-    }
-}
-
-@Composable
-private fun CompletingContent(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier
-            .fillMaxWidth()
-            .background(color = RadixTheme.colors.defaultBackground)
-            .padding(RadixTheme.dimensions.paddingLarge),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(RadixTheme.dimensions.paddingDefault)
-    ) {
-        Image(
-            painter = painterResource(
-                id = com.babylon.wallet.android.designsystem.R.drawable.check_circle_outline
-            ),
-            alpha = 0.2F,
-            contentDescription = null
-        )
-        Text(
-            text = stringResource(R.string.transactionStatus_completing_text),
-            style = RadixTheme.typography.body1Regular,
-            color = RadixTheme.colors.gray1
-        )
-        Spacer(Modifier.height(36.dp))
-    }
-}
-
-internal class MobileConnectParameterProvider : PreviewParameterProvider<Boolean> {
-    override val values: Sequence<Boolean> = sequenceOf(true, false)
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SuccessBottomDialogPreview(
-    @PreviewParameter(MobileConnectParameterProvider::class) isMobileConnect: Boolean
-) {
-    RadixWalletTheme {
-        SuccessContent(
-            transactionAddress = "txid_tdx_21_1nsdfruuw5gd6tsh07ur5mgq4tjpns9vxj0nnaahaxpxmxapjzrfqmfzr4s",
-            isMobileConnect = isMobileConnect,
-            isInternal = false,
-            dAppName = null
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun CompletingBottomDialogPreview() {
-    RadixWalletTheme {
-        CompletingContent()
-    }
 }
